Refactor ToDoList entity and add creation endpoint
Changed the 'items' property in ToDoList from val to var, allowing it to be nullable and initialized as an empty list (because spring maps an empty array to a null value???). Also, added a new 'createToDoList' endpoint in ToDoListController, enabling the creation of a new ToDoList. This change allows more flexibility in handling 'items' and enhances the feature set by allowing the creation of ToDo lists.